Intermediate ML-- Analyzing wine type and quality
Analyzing wine types and wine quality using dataset pertaining to wine
samples with a focus on type and quality ratings. Special emphasis is on exploratory data analysis, which
is often overlooked by data scientists in a hurry to build and deploy models. We will look at multiple ways to analyze and visualize our data and its features, including descriptive and
inferential statistics and univariate and multivariate analysis. Various aspects of model training, prediction, evaluation, tuning,
and interpretation are covered in detail.